+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### 7.0.0-beta.3 / TBD
+* Added support for the free-busy endpoint
+* Fix `getAvailability` response type
+* Fix path for destroying a redirect URI
+* Fix model for updating an Event
+
 ### 7.0.0-beta.2 / 2023-08-21
 * Fix issue with getting the current Nylas SDK package version
 * Fix issue with ESM imports not working correctly